4.1.1.1637
Release Notes

Release 4.1.1 features
Ellucian Mobile 4.1.1 is a mobile device software release only, supporting new features available in iOS 9 and Android 6.0. There are no changes to the Ellucian Mobile Sever; therefore, it is not included in the 4.1.1 software distribution. If you are a new customer implementing Ellucian Mobile for the first time, you’ll want to download the 4.0 software package to obtain the Mobile Server.
In this release, we have dropped support for iOS 7 to take advantage of many new iOS technologies that improve accessibility including better right to left support, Dynamic Type, and other enhancements.

iOS9 Support
The Audio/Video Module now plays content in the background and supports the new Picture in Picture feature available in iOS 9 on eligible iPad models.

watchOS 2 Support
Ellucian Mobile 4.1.1 now supports watchOS 2, and several user interaction improvements have been implemented providing a better overall Apple Watch experience. This and later releases will not support watchOS1.

Android 6.0
The 4.1.1 version of Ellucian GO has also gone through a UI modernization process providing a new look for many areas of the application.
Ellucian GO now supports a new Android Permission Model, which only prompts the user for access to their device location if the Maps Module is launched by the user. Previously this permission was requested at install time even if the Maps Module wasn’t being utilized.

Resolved Change Requests
CR-000125648 - Platform Developer Guide contains an incorrect link to Android Studio. - Corrected the Documentation
CR-000121956 - Android application crash during login. - Fixed the crashing behavior.
CR-000131750 - Horizontal Line Menu button hard to see on Android. - UI Enhancements addressed this issue.
CR-000111756 - Ellucian Go Android App crashes on close. - Fixed the crashing behavior.
CR-000111861 - Android application User is logged out of the app if it crashes. - Fixed the crashing behavior.
CR-000125703 - Android version of Ellucian GO showing HTML markup in notifications. - Notifications on android now support html content.
CR-000104598 - Having the wrong number of characters for a branding color code crashes the app in Android. - Cloud Configuration server no longer allows invalid color values.
CR-000123090 - Mobile 3.7 iOS App crashes in landscape on iOS 8.1 when launched from test URL. - Fixed the crashing behavior in later releases.
CR-000123174 - iPhone 6 screen resizing. Bug in the Directory details screen. - Fixed with iOS SDK updates.
CR-000126106 - Events Module on Android does not handle HTML content. - Events now display HTML content properly.
CR-000129461 - Event Detail doesn't display line breaks correctly. - Events now display HTML content properly.
CR-000132054 - Ellucian Mobile Not Compatible for installation with Android 5.1.1. - Ellucian Mobile now supports Android 4.x-6.x.
CR-000106661 - Duplicate content Description for list Image style causes Android Studio compilation failure. - Removed duplicate content.
CR-000111420 - Important Numbers open- close icon not positioned properly on Android. - Important Numbers Module has been redesigned.
CR-000111535 - Change Add to Calendar label to an icon on Android Events detail. - An icon/button has replaced the "Add to calendar" text.
CR-000111718 - News and Events detail screens need indicator there are more items on iPad in portrait orientation. - The list of events is now always displayed alongside the details.
CR-000123200 - Selection arrow displays on wrong side of News and Events on Android when in Arabic. - Corrected the RTL error.
CR-000130165 - iOS News/Events: Improved text wrapping is needed for iPhone 6 and 6 Plus. - Resolved text wrapping.
CR-000130218 - iOS: Accessibility - Menu button is being read as "button". - Now is read as "Menu Button".
CR-000130220 - iOS: Accessibility - When the menu tray is expanded, normal navigation doesn't function properly. - You can now use the rotator to navigate the open menu.
CR-000133948 - Use a WebView for Android Feed Detail Instead of TextView. - Feeds now display HTML content properly.
CR-000133950 - Provide a Term selection for classes/grades on Android. - Provided a term selection button.
CR-000111364 - Need visual indicator on Notification detail that there are more notifications to see on iPad. - The list of notifications is now always displayed alongside the details.
CR-000123199 - Android: Notification description doesn't display HTML content. - Notifications now display HTML content properly.
CR-000133953 - Display News and Event list for iPad in portrait and landscape orientations. - The list of notifications is now always displayed alongside the details.
CR-000133954 - Remove the 'Ellucian Mobile' target from iOS. - Ellucian specific target has been removed.
CR-000133955 - Login Spinner not visible with some background colors. - Changed how colors are handled at login.
CR-000133956 - Update Platform Edition developer guide to include the correct customization steps. - Platform Developer's Guide was updated.
CR-000133957 - Remove unnecessary files from the Android code. - Unnecessary files removed. Extensions are now displayed properly.
CR-000133958 - Emergency Numbers with extensions are not properly listed. - If the school selects a custom header icon from the Cloud Configuration tool, it will be displayed next to the header title.
CR-000103637 - For Android app only. The default menu icon and the Ellucian "e" icon are displayed at the same time. - If the school uses the default menu icon, the Ellucian "e" icon will no longer be displayed.

+++ iOS/Assignments Today/TodayViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import NotificationCenter
 
-class TodayViewController: UITableViewController, NCWidgetProviding, UITableViewDataSource, UITableViewDelegate {
+class TodayViewController: UITableViewController, NCWidgetProviding {
     
     var items : [NSDictionary]?
     var noILP : Bool = false
@@ -26,16 +26,16 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
         let defaults = AppGroupUtilities.userDefaults()
         items = defaults?.objectForKey("today-widget-assignments") as! [NSDictionary]?
         self.reloadTable()
-        
-        let url = defaults?.objectForKey("ilp-url") as! NSString?
+
         self.disconnected = UserInfo.userauth() == nil
         
         addButtonToDisconnectedFooter()
-        fetch()
 
     }
     
-    func widgetPerformUpdateWithCompletionHandler(completionHandler: ((NCUpdateResult) -> Void)!) {
+    func widgetPerformUpdateWithCompletionHandler(completionHandler: ((NCUpdateResult) -> Void)) {
+        fetch()
+        
         completionHandler(NCUpdateResult.NewData)
     }
     
@@ -47,8 +47,8 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
     }
     
     override func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        var itemNumber = indexPath.row
-        let cell : UITableViewCell = tableView.dequeueReusableCellWithIdentifier("Assignment Today Cell", forIndexPath: indexPath) as! UITableViewCell
+        let itemNumber = indexPath.row
+        let cell : UITableViewCell = tableView.dequeueReusableCellWithIdentifier("Assignment Today Cell", forIndexPath: indexPath) as UITableViewCell
         
         if let items = self.items {
             let item = items[itemNumber]
@@ -61,7 +61,7 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
     }
     
     override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
-        var itemNumber = indexPath.row
+        let itemNumber = indexPath.row
         if let items = self.items {
             let item = items[itemNumber]
             let itemUrl = item["url"] as! String
@@ -71,25 +71,7 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
             self.extensionContext!.openURL(url!, completionHandler: nil)
         }
     }
-    
-    override func viewWillTransitionToSize(size: CGSize, withTransitionCoordinator coordinator: UIViewControllerTransitionCoordinator)
-    {
-        super.viewWillTransitionToSize(size, withTransitionCoordinator: coordinator)
-        
-        if let safeCoordinator = coordinator as UIViewControllerTransitionCoordinator?
-        {
-            println("coordinator != nil")
-            safeCoordinator.animateAlongsideTransition({ context in
-                self.tableView.frame = CGRectMake(0, 0, size.width, size.height)
-                }, completion: nil)
 
-        }
-        else
-        {
-            println("coordinator == nil")
-        }
-    }
-    
     func reload() {
         self.tableView.reloadData()
         self.preferredContentSize = self.tableView.contentSize
@@ -97,15 +79,15 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
     }
     
     func filterItems(items: [NSDictionary]) -> [NSDictionary]? {
-        println("Start filtering")
+        print("Start filtering")
         let sourceToDateFormatter = NSDateFormatter()
         sourceToDateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss'Z'"
         sourceToDateFormatter.timeZone = NSTimeZone(forSecondsFromGMT: 0)
         
-        var comparisonFormatter = NSDateFormatter()
+        let comparisonFormatter = NSDateFormatter()
         comparisonFormatter.dateFormat = "yyyy-MM-dd"
         comparisonFormatter.timeZone = NSTimeZone.localTimeZone()
-        var dateNow = NSDate()
+        let dateNow = NSDate()
         
         let dateNowString = comparisonFormatter.stringFromDate(dateNow)
         
@@ -118,7 +100,7 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
                 return false
             }
         }
-        println("End filtering")
+        print("End filtering")
         return filteredArray
     }
     
@@ -178,7 +160,7 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
         CGContextAddPath(ctx, path.CGPath);
         CGContextFillPath(ctx);
         
-        CGContextSetBlendMode(ctx, kCGBlendModeDestinationOut);
+        CGContextSetBlendMode(ctx, CGBlendMode.DestinationOut);
         let center = CGPointMake((size.width - textSize.width) / 2.0, (size.height - textSize.height) / 2.0);
         string.drawAtPoint(center, withAttributes: textAttributes)
         
@@ -209,7 +191,7 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
     }
     
     private func addButtonToDisconnectedFooter() {
-        let button = UIButton.buttonWithType(.System) as! UIButton
+        let button = UIButton(type: .System) as UIButton
         button.frame = CGRectMake(0, 0, disconnectedButtonView.frame.width, disconnectedButtonView.frame.height)
         button.contentHorizontalAlignment = .Center
         button.backgroundColor = UIColor.clearColor()
@@ -221,7 +203,7 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
         button.setImage(imageCutout, forState: UIControlState.Highlighted)
         button.addTarget(self, action: "signIn:", forControlEvents: .TouchUpInside)
         
-        var visualEffectView = UIVisualEffectView(effect: UIVibrancyEffect.notificationCenterVibrancyEffect())
+        let visualEffectView = UIVisualEffectView(effect: UIVibrancyEffect.notificationCenterVibrancyEffect())
         visualEffectView.frame = CGRectMake(0, 0, disconnectedButtonView.frame.size.width, disconnectedButtonView.frame.height)
         visualEffectView.contentView.addSubview(button)
         
@@ -231,41 +213,43 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
     func fetch() {
         let defaults = AppGroupUtilities.userDefaults()
         let url = defaults?.objectForKey("ilp-url") as! NSString?
-        println("Assignments Today widgetPerformUpdateWithCompletionHandler")
+        print("Assignments Today widgetPerformUpdateWithCompletionHandler")
         
         items = defaults?.objectForKey("today-widget-assignments") as! [NSDictionary]?
         
-        var storage : NSHTTPCookieStorage = NSHTTPCookieStorage.sharedHTTPCookieStorage()
-        for cookie in storage.cookies as! [NSHTTPCookie]{
-            storage.deleteCookie(cookie)
+        let storage : NSHTTPCookieStorage = NSHTTPCookieStorage.sharedHTTPCookieStorage()
+        if let cookies = storage.cookies {
+            for cookie in cookies {
+                storage.deleteCookie(cookie)
+            }
         }
 
-        var cookiesArray : NSArray? = defaults?.objectForKey("cookieArray") as! NSArray?
+        let cookiesArray : NSArray? = defaults?.objectForKey("cookieArray") as! NSArray?
         if let cookiesArray = cookiesArray {
             for cookieItem in cookiesArray  {
-                let cookieDictionary = cookieItem as! [NSObject : AnyObject]
+                let cookieDictionary = cookieItem as! [String : AnyObject]
                 let cookie = NSHTTPCookie(properties: cookieDictionary)
                 NSHTTPCookieStorage.sharedHTTPCookieStorage().setCookie(cookie!)
             }
         }
 
         if let ilpUrl = url {
-            println("Assignments Today url: \(ilpUrl)")
+            print("Assignments Today url: \(ilpUrl)")
             if let username = UserInfo.userauth() {
-                println("Assignments Today has username")
+                print("Assignments Today has username")
                 let config = NSURLSessionConfiguration.defaultSessionConfiguration()
                 let password = UserInfo.password()
                 if let password = password {
                     let userPasswordString = "\(username):\(password)"
                     let userPasswordData = userPasswordString.dataUsingEncoding(NSUTF8StringEncoding)
-                    let base64EncodedCredential = userPasswordData!.base64EncodedStringWithOptions(nil)
+                    let base64EncodedCredential = userPasswordData!.base64EncodedStringWithOptions([])
                     let authString = "Basic \(base64EncodedCredential)"
                     config.HTTPAdditionalHeaders = ["Authorization" : authString]
                 }
                 let session = NSURLSession(configuration: config)
                 
                 if let studentId = UserInfo.userid() {
-                    println("Assignments Today has studentId")
+                    print("Assignments Today has studentId")
                     let fullUrl = NSURL(string: "\(ilpUrl)/\(studentId)/assignments")!
                     
                     
@@ -273,20 +257,20 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
                         (let data, let response, let error) in
                         
                         if let httpRes = response as? NSHTTPURLResponse {
-                            println("Assignments Today response code: \(httpRes.statusCode)")
+                            print("Assignments Today response code: \(httpRes.statusCode)")
                             if httpRes.statusCode == 200 {
-                                let json = JSON(data: data)
+                                let json = JSON(data: data!)
                                 
                                 
                                 if(json == nil) {
-                                    println("Assignments Today disconnected")
+                                    print("Assignments Today disconnected")
                                     self.disconnected = true
                                     self.reloadTable()
                                 } else {
                                     var items = [NSDictionary]()
                                     let assignmentList: Array<JSON> = json["assignments"].arrayValue
                                     
-                                    for assignmentJson in json["assignments"].arrayValue {
+                                    for assignmentJson in assignmentList {
                                         items.append(["sectionId" : assignmentJson["sectionId"].stringValue,
                                             "courseName": assignmentJson["courseName"].stringValue, "courseSectionNumber": assignmentJson["courseSectionNumber"].stringValue, "name": assignmentJson["name"].stringValue, "assignmentDescription": assignmentJson["description"].stringValue, "dueDate": assignmentJson["dueDate"].stringValue, "url": assignmentJson["url"].stringValue])
                                     }
@@ -295,35 +279,35 @@ class TodayViewController: UITableViewController, NCWidgetProviding, UITableView
                                     
                                     self.reloadTable()
                                     defaults?.setObject(self.items, forKey: "today-widget-assignments")
-                                    println("Assignments Today count: \(self.items!.count)")
+                                    print("Assignments Today count: \(self.items!.count)")
                                 }
                             } else {
                                 self.disconnected = true
                                 self.reloadTable()
                             }
                         } else {
-                            println("error \(error)")
+                            print("error \(error)")
                         }
                     }
                     task.resume()
                 }
             } else {
-                println("Assignments Today disconnected")
+                print("Assignments Today disconnected")
                 self.disconnected = true
                 reloadTable()
             }
         } else {
-            println("Assignments Today no ILP")
+            print("Assignments Today no ILP")
             self.noILP = true
             reloadTable()
         }
     }
     
     func reloadTable() {
-        println("Start reload")
+        print("Start reload")
         dispatch_async(dispatch_get_main_queue()) {
             self.reload()
-            println("End reload")
+            print("End reload")
         }
     }
 }
